// src/modules/Publisher/publisher.interface.ts
import { Types } from 'mongoose';
export type TGender = 'male' | 'female' | 'other';
export type TBloodGroup =
  | 'A+'
  | 'A-'
  | 'B+'
  | 'B-'
  | 'AB+'
  | 'AB-'
  | 'O+'
  | 'O-';
export type TUserName = {
  firstName: string;
  middleName?: string;
  lastName: string;
};
export type TGuardian = {
  fatherName: string;
  fatherOccupation: string;
  fatherContactNo: string;
  motherName: string;
  motherOccupation: string;
  motherContactNo: string;
};

export type TPublisher = {
  id: string;
  user: Types.ObjectId;
  name: TUserName;
  bio?: string;
  gender: TGender;
  dateOfBirth?: Date;
  email: string;
  contactNo: string;
  emergencyContactNo: string;
  guardian: TGuardian;
  bloodGroup?: TBloodGroup;
  presentAddress: string;
  permanentAddress: string;
  profileImage: string;
  experience: string;
  isFreelancer: boolean;
  department: Types.ObjectId;
  seasonalDate: Types.ObjectId;
  isDeleted: boolean;
};